beware next iterations (Closes: #6216).
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/release-6.0/work/pkgs@22769 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ untangle-gateway/files/usr/share/untangle-gateway/knoppix2lenny.sh
@@ -273,7 +273,9 @@ stepRemoveUnwantedPackaged() {
 
   # this removes X 3.3 packages (and some untangle-* packages, see above)
   removePackagesRegex "(xserver|xbase-clients).*3\.3\.6"
-  rm -fr /usr/X11R6/bin/*
+  if [ ! -L /usr/X11R6/bin ] ; then 
+    rm -fr /usr/X11R6/bin/*
+  fi
 
   removePackages "xfce4-mixer-lib-alsa"
